conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20211102.0=hc377ac9_0
aiohttp=3.8.3=py310h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
arrow-cpp=11.0.0=hc7aafb3_1
async-timeout=4.0.2=py310hca03da5_0
attrs=22.1.0=py310hca03da5_0
aws-c-common=0.6.8=h80987f9_1
aws-c-event-stream=0.1.6=h313beb8_6
aws-checksums=0.1.11=h80987f9_2
aws-sdk-cpp=1.8.185=ha71a6ea_1
blas=1.0=openblas
boost-cpp=1.73.0=h1a28f6b_12
bottleneck=1.3.5=py310h96f19d2_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotlipy=0.7.0=py310h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.0=h80987f9_0
ca-certificates=2023.05.30=hca03da5_0
certifi=2023.5.7=py310hca03da5_0
cffi=1.15.1=py310h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310hca03da5_0
contourpy=1.0.5=py310h525c30c_0
cryptography=39.0.1=py310h834c97f_2
cycler=0.11.0=pyhd3eb1b0_0
datasets=2.12.0=py310hca03da5_0
dill=0.3.6=py310hca03da5_0
filelock=3.9.0=py310hca03da5_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=py310h80987f9_0
fsspec=2023.4.0=py310hca03da5_0
gflags=2.2.2=hc377ac9_0
giflib=5.2.1=h80987f9_3
glog=0.5.0=hc377ac9_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py310h8c48613_0
grpc-cpp=1.48.2=hc60591f_1
huggingface_accelerate=0.20.3=py310hca03da5_0
huggingface_hub=0.15.1=py310hca03da5_0
icu=68.1=hc377ac9_0
idna=3.4=py310hca03da5_0
jinja2=3.1.2=py310hca03da5_0
joblib=1.2.0=py310hca03da5_0
jpeg=9e=h80987f9_1
kiwisolver=1.4.4=py310h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libboost=1.73.0=h49e8a49_12
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcurl=8.1.1=h3e2b118_1
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libnghttp2=1.52.0=h62f6fdd_1
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libssh2=1.10.0=h02f6b3c_2
libthrift=0.15.0=h73c2103_2
libtiff=4.5.0=h313beb8_2
libuv=1.44.2=h80987f9_0
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
lightning-utilities=0.7.1=py310hca03da5_1
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py310h1a28f6b_0
matplotlib-base=3.7.1=py310h46d7db6_1
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.2.1=pypi_0
multidict=6.0.2=py310h1a28f6b_0
multiprocess=0.70.14=py310hca03da5_0
munkres=1.1.4=py_0
ncurses=6.4=h313beb8_0
networkx=2.8.4=py310hca03da5_1
ninja=1.10.2=hca03da5_5
ninja-base=1.10.2=h525c30c_5
numexpr=2.8.4=py310hecc3335_1
numpy=1.25.0=py310h3b2db8e_0
numpy-base=1.25.0=py310ha9811e2_0
openssl=3.0.9=h1a28f6b_0
orc=1.7.4=hdca1487_1
packaging=23.0=py310hca03da5_0
pandas=1.5.3=py310h46d7db6_0
pillow=9.4.0=py310h313beb8_0
pip=23.1.2=py310hca03da5_0
pooch=1.4.0=pyhd3eb1b0_0
protobuf=3.20.3=py310h313beb8_0
psutil=5.9.0=py310h1a28f6b_0
pyarrow=11.0.0=py310hbfed03b_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py310hca03da5_0
pyparsing=3.0.9=py310hca03da5_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.11=hb885b13_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-xxhash=2.0.2=py310h1a28f6b_1
pytorch=2.0.1=gpu_mps_py310h20d1048_0
pytorch-lightning=2.0.3=py310hca03da5_0
pytz=2022.7=py310hca03da5_0
pyyaml=6.0=py310h80987f9_1
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
regex=2022.7.9=py310h1a28f6b_0
requests=2.29.0=py310hca03da5_0
responses=0.13.3=pyhd3eb1b0_0
sacremoses=0.0.43=pyhd3eb1b0_0
scikit-learn=1.2.2=py310h313beb8_1
scipy=1.10.1=py310h20cbe94_1
setuptools=67.8.0=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h80987f9_2
snappy=1.1.9=hc377ac9_0
sqlite=3.41.2=h80987f9_0
sympy=1.11.1=py310hca03da5_0
tensorboardx=2.2=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.13.2=py310h3dd52b7_1
torch-lr-finder=0.2.1=pyhd3eb1b0_0
torchmetrics=0.11.4=py310h33ce5c2_1
tqdm=4.65.0=py310h33ce5c2_0
transformers=4.29.2=py310hca03da5_0
typing-extensions=4.6.3=py310hca03da5_0
typing_extensions=4.6.3=py310hca03da5_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py310hca03da5_0
utf8proc=2.6.1=h1a28f6b_0
wheel=0.38.4=py310hca03da5_0
xxhash=0.8.0=h1a28f6b_3
xz=5.4.2=h80987f9_0
yaml=0.2.5=h1a28f6b_0
yarl=1.8.1=py310h1a28f6b_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
